Prevent curator jobs from failing if nothing to do
This prevents curator from failing if there is no work to do, such as no indicies or no indicies to delete. This job would error out for 30 days without this setting. This prevents an error code of -1 if there's nothing to do. It still logs out what action it took.